put managed property listeners back on beanstub
