+++ community-modules/core/src/context/beanStub.ts
@@ -1,6 +1,13 @@
+import type { GridOptions } from '../entities/gridOptions';
 import type { EventService } from '../eventService';
 import type { AgEvent, AgEventListener } from '../events';
-import type { GridOptionsService } from '../gridOptionsService';
+import type {
+    GridOptionsService,
+    PropertyChangedEvent,
+    PropertyChangedListener,
+    PropertyValueChangedEvent,
+    PropertyValueChangedListener,
+} from '../gridOptionsService';
 import type { IEventEmitter } from '../interfaces/iEventEmitter';
 import type { IFrameworkOverrides } from '../interfaces/iFrameworkOverrides';
 import { LocalEventService } from '../localEventService';
@@ -66,6 +73,84 @@ export abstract class BeanStub implements Bean, IEventEmitter {
         this.destroyed = true;
 
         this.dispatchEvent({ type: BeanStub.EVENT_DESTROYED });
+    }
+
+    /**
+     * Setup a managed property listener for the given GridOption property.
+     * However, stores the destroy function in the beanStub so that if this bean
+     * is a component the destroy function will be called when the component is destroyed
+     * as opposed to being cleaned up only when the GridOptionsService is destroyed.
+     */
+    private setupGridOptionListener<K extends keyof GridOptions>(
+        event: keyof GridOptions,
+        listener: PropertyValueChangedListener<K>
+    ): () => null {
+        this.gos.addPropertyEventListener(event, listener);
+        const destroyFunc: () => null = () => {
+            this.gos.removePropertyEventListener(event, listener);
+            return null;
+        };
+        this.addDestroyFunc(destroyFunc);
+
+        return destroyFunc;
+    }
+
+    /**
+     * Setup a managed property listener for the given GridOption property.
+     * @param event GridOption property to listen to changes for.
+     * @param listener Listener to run when property value changes
+     */
+    public addManagedPropertyListener<K extends keyof GridOptions>(
+        event: K,
+        listener: PropertyValueChangedListener<K>
+    ): () => null {
+        if (!this.isAlive()) {
+            return () => null;
+        }
+
+        return this.setupGridOptionListener(event, listener);
+    }
+
+    private propertyListenerId = 0;
+    // Enable multiple grid properties to be updated together by the user but only trigger shared logic once.
+    // Closely related to logic in ComponentUtil.ts
+    private lastChangeSetIdLookup: Record<string, number> = {};
+    /**
+     * Setup managed property listeners for the given set of GridOption properties.
+     * The listener will be run if any of the property changes but will only run once if
+     * multiple of the properties change within the same framework lifecycle event.
+     * Works on the basis that GridOptionsService updates all properties *before* any property change events are fired.
+     * @param events Array of GridOption properties to listen for changes too.
+     * @param listener Shared listener to run if any of the properties change
+     */
+    public addManagedPropertyListeners(events: (keyof GridOptions)[], listener: PropertyChangedListener): void {
+        if (!this.isAlive()) {
+            return;
+        }
+
+        // Ensure each set of events can run for the same changeSetId
+        const eventsKey = events.join('-') + this.propertyListenerId++;
+
+        const wrappedListener = (event: PropertyValueChangedEvent<any>) => {
+            if (event.changeSet) {
+                // ChangeSet is only set when the property change is part of a group of changes from ComponentUtils
+                // Direct api calls should always be run as
+                if (event.changeSet && event.changeSet.id === this.lastChangeSetIdLookup[eventsKey]) {
+                    // Already run the listener for this set of prop changes so don't run again
+                    return;
+                }
+                this.lastChangeSetIdLookup[eventsKey] = event.changeSet.id;
+            }
+            // Don't expose the underlying event value changes to the group listener.
+            const propertiesChangeEvent: PropertyChangedEvent = {
+                type: 'gridPropertyChanged',
+                changeSet: event.changeSet,
+                source: event.source,
+            };
+            listener(propertiesChangeEvent);
+        };
+
+        events.forEach((event) => this.setupGridOptionListener(event, wrappedListener));
     }
 
     public addEventListener(eventType: string, listener: AgEventListener): void {
